--- VendingMachineApp/VendingMachineApp/Models/Tags.swift
@@ -1,21 +1,0 @@
-//
-//  Tags.swift
-//  VendingMachineApp
-//
-//  Created by Chaewan Park on 2020/03/11.
-//  Copyright © 2020 Chaewan Park. All rights reserved.
-//
-
-import Foundation
-
-class Tags {
-    enum Label: Int {
-        case coke = 101
-        case fanta, strawberryMilk, chocolateMilk, georgia, top
-    }
-    
-    enum Button: Int {
-        case coke = 201
-        case fanta, strawberryMilk, chocolateMilk, georgia, top
-    }
-}
